requirements.txt
# --------- pytorch --------- #
torch==1.13.1
pytorch-lightning==1.9.0
torchmetrics>=0.3.2
onnxruntime
onnxruntime-gpu
onnx

# --------- hydra --------- #
hydra-core==1.3.1
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.3.0.dev0
# hydra-ax-sweeper==1.1.0
# hydra-ray-launcher==0.1.2
# hydra-submitit-launcher==1.1.0

# --------- loggers --------- #
wandb>=0.10.30
# neptune-client
# mlflow
# comet-ml
# torch_tb_profiler

# --------- linters --------- #
pre-commit
black
isort
flake8

# --------- others --------- #
jupyterlab
python-dotenv
rich
pytest
sh
scikit-learn
seaborn
pudb
# dvc


# --------- kuielab --------- #
aicrowd_api
coloredlogs
loguru
openunmix
musdb
asteroid>=0.5.0
demucs
numpy==1.24.1
tensorboard
pyrootutils